Fixed: cast dwScrollPastEOF to int.
git-svn-id: https://svn.code.sf.net/p/akelpad/codesvn/trunk@4512 24cdac72-59cb-4434-bdd7-86cdcaa6fc6e

+++ akelpad_4/AkelEdit/AkelEdit.c
@@ -7255,7 +7255,7 @@ void AE_UpdateVScrollMax(AKELEDIT *ae, BOOL bTextUpdate)
       dwScrollPastEOF=ae->rcDraw.bottom - ae->rcDraw.top;
     else
       dwScrollPastEOF=ae->popt->dwScrollPastEOF * (ae->rcDraw.bottom - ae->rcDraw.top) / 100;
-    ae->ptxt->nVScrollMax+=max(dwScrollPastEOF - ae->ptxt->nCharHeight - 1, 0);
+    ae->ptxt->nVScrollMax+=max((int)dwScrollPastEOF - ae->ptxt->nCharHeight - 1, 0);
   }
 }
